services/fetch_news_crypto/requirements.txt
fastapi
uvicorn
feedparser
newspaper3k
lxml_html_clean
